lambda_functions/export_rds_to_s3/requirements.txt
aws-psycopg2
boto3
pandas
s3fs
fsspec
